+++ src/transformers/models/granite/attention/__init__.py
@@ -1,8 +1,3 @@
-import inspect
-from typing import Tuple
-
-import torch
-
 from ..config import GraniteConfig
 from ..enums import AttentionHeadType, AttentionImplementation, PositionEmbeddingType
 from .base import Attention
@@ -10,14 +5,6 @@ from .flash import FlashAttention
 from .math import MathAttention
 from .padding_free import PaddingFreeAttention
 from .sdpa import SDPA
-from .utils import (
-    interleave_query_key_value_tensor_for_gqa,
-    interleave_query_key_value_tensor_for_mha,
-    interleave_query_key_value_tensor_for_mqa,
-    split_query_key_value_tensor_for_gqa,
-    split_query_key_value_tensor_for_mha,
-    split_query_key_value_tensor_for_mqa,
-)
 
 
 _ATTENTION_MODULES = {
@@ -26,20 +13,6 @@ _ATTENTION_MODULES = {
     AttentionImplementation.sdpa.value: SDPA,
     AttentionImplementation.flash.value: FlashAttention,
     AttentionImplementation.padding_free.value: PaddingFreeAttention,
-}
-
-
-_INTERLEAVE_FUNCTIONS = {
-    AttentionHeadType.mha.value: interleave_query_key_value_tensor_for_mha,
-    AttentionHeadType.mqa.value: interleave_query_key_value_tensor_for_mqa,
-    AttentionHeadType.gqa.value: interleave_query_key_value_tensor_for_gqa,
-}
-
-
-_SPLIT_FUNCTIONS = {
-    AttentionHeadType.mha.value: split_query_key_value_tensor_for_mha,
-    AttentionHeadType.mqa.value: split_query_key_value_tensor_for_mqa,
-    AttentionHeadType.gqa.value: split_query_key_value_tensor_for_gqa,
 }
 
 
@@ -69,49 +42,3 @@ def get_attention_module(
         )
 
     raise ValueError(f"unexpected `attention_implementation` {attention_implementaton}")
-
-
-def interleave_query_key_value_tensor_for_attention(
-    query_weight: torch.Tensor,
-    key_weight: torch.Tensor,
-    value_weight: torch.Tensor,
-    num_heads: int,
-    num_key_value_heads: int,
-    head_dim: int,
-    attention_head_type: AttentionHeadType,
-) -> torch.Tensor:
-    if attention_head_type.value in _INTERLEAVE_FUNCTIONS:
-        interleave_function = _INTERLEAVE_FUNCTIONS[attention_head_type.value]
-        interleave_function_parameters = inspect.signature(interleave_function).parameters.keys()
-
-        parameters_to_pass = {}
-        this_function_parameters = locals()
-        for parameter in interleave_function_parameters:
-            parameters_to_pass[parameter] = this_function_parameters[parameter]
-
-        query_key_value_weight = interleave_function(**parameters_to_pass)
-
-        return query_key_value_weight
-
-    raise ValueError(f"unexpected `attention_head_type` {attention_head_type}")
-
-
-def split_query_key_value_tensor_for_attention(
-    query_key_value_weight: torch.Tensor,
-    num_heads: int,
-    num_key_value_heads: int,
-    head_dim: int,
-    attention_head_type: AttentionHeadType,
-) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    if attention_head_type.value in _SPLIT_FUNCTIONS:
-        split_function = _SPLIT_FUNCTIONS[attention_head_type.value]
-        split_function_parameters = inspect.signature(split_function).parameters.keys()
-
-        parameters_to_pass = {}
-        this_function_parameters = locals()
-        for parameter in split_function_parameters:
-            parameters_to_pass[parameter] = this_function_parameters[parameter]
-
-        return split_function(**parameters_to_pass)
-
-    raise ValueError(f"unexpected `attention_head_type` {attention_head_type}")
